Display field's function `with` moved to the Field class for using it in each form field.
